Make possible to provide custom OPF and NCX/HTML toc templates.

+++ lib/index.coffee
@@ -39,6 +39,9 @@ class EPub
       appendChapterTitles: true
       date: new Date().toISOString()
       lang: "en"
+      customOpfTemplatePath: null
+      customNcxTocTemplatePath: null
+      customHtmlTocTemplatePath: null
     }, options
 
     if _.isString @options.author
@@ -94,6 +97,10 @@ class EPub
             self.defer.reject(err)
         , (err)->
           self.defer.reject(err)
+      , (err)->
+        self.defer.reject(err)
+    , (err)->
+      self.defer.reject(err)
 
   generateTempFile: ()->
     generateDefer = new Q.defer()
@@ -129,10 +136,25 @@ class EPub
     fs.mkdirSync(@uuid + "/META-INF")
     fs.writeFileSync( "#{@uuid}/META-INF/container.xml", """<?xml version="1.0" encoding="UTF-8" ?><container version="1.0" xmlns="urn:oasis:names:tc:opendocument:xmlns:container"><rootfiles><rootfile full-path="OEBPS/content.opf" media-type="application/oebps-package+xml"/></rootfiles></container>""")
 
+    opfPath = self.options.customOpfTemplatePath or path.resolve(__dirname, "./content.ejs")
+    if !fs.existsSync(opfPath)
+      generateDefer.reject(new Error('Custom file to OPF template not found.'))
+      return generateDefer.promise
+
+    ncxTocPath = self.options.customNcxTocTemplatePath or path.resolve(__dirname , "./toc.ejs" )
+    if !fs.existsSync(ncxTocPath)
+      generateDefer.reject(new Error('Custom file the NCX toc template not found.'))
+      return generateDefer.promise
+
+    htmlTocPath = self.options.customHtmlTocTemplatePath or path.resolve(__dirname, "./content.html")
+    if !fs.existsSync(htmlTocPath)
+      generateDefer.reject(new Error('Custom file to HTML toc template not found.'))
+      return generateDefer.promise
+
     Q.all([
-      Q.nfcall ejs.renderFile, path.resolve(__dirname, "./content.ejs"), self.options
-      Q.nfcall ejs.renderFile, path.resolve( __dirname , "./toc.ejs" ), self.options
-      Q.nfcall ejs.renderFile, path.resolve(__dirname, "./content.html"), self.options
+      Q.nfcall ejs.renderFile, opfPath, self.options
+      Q.nfcall ejs.renderFile, ncxTocPath, self.options
+      Q.nfcall ejs.renderFile, htmlTocPath, self.options
     ]).spread (data1, data2, data3)->
       fs.writeFileSync(path.resolve(self.uuid , "./OEBPS/content.opf"), data1)
       fs.writeFileSync(path.resolve(self.uuid , "./OEBPS/toc.ncx"), data2)
